Add a new user defined calendar type.
A new calendar type "user_defined" is now available. This allows the users to create a custom calendar that we can configured to be suitable for planets other than the Earth.

An user defined calendar is always defined by two mandatory attributes:
    - `day_length`: the duration of a day, in seconds
    - and either:
        - `month_length`: an array containing the duration of each month, in days (the number of elements in the array is the number of months in a year)
        - or `year_length`: the duration of a year, in seconds (in that case, the calendar does not have months).

If the calendar has months (i.e. `month_length` attribute is set) and only in that case, it is possible to define leap years in order to compensate for the duration of an astronomical year not being a multiple of the day length. The leap years are defined by two mandatory attributes:
    - `leap_year_month`: the month to which the extra day will be added in case of leap year, expressed as an integer number in the range [1, numberOfMonths]
    - and `leap_year_drift`: the fraction of a day representing the yearly drift between the calendar year and the astronomical year, expressed as a real number in the range [0, 1).
Optionally, one can define the `leap_year_drift_offset` attribute to set the original drift at the beginning of the time origin's year, again expressed as a real number in the range [0, 1). If `leap_year_drift_offset + leap_year_drift` is greater or equal to 1, then the first year will be a leap year.

For example, the following configuration creates a Gregorian-like calendar:
    `<calendar type="user_defined" start_date="2012-03-01 15:00:00" time_origin="2012-02-28 15:00:00 + 1d" day_length="86400" month_lengths="(1, 12) [31 28 31 30 31 30 31 31 30 31 30 31]" leap_year_month="2" leap_year_drift="0.25" leap_year_drift_offset="0.75" />`

Note that dates attributes must be written differently in the configuration file when using an user defined calendar without months:
    - if the year length is greater than the day length, the input format is `year-day hh:min:sec` instead of `year-month-day hh:min:sec`
    - if the day length is greater or equal to the year length, the input format is `year hh:min:sec`.
In all cases, it is still possible to use the `date + duration` notation to build a date (with both the date and duration parts being optional).

The Fortran interface has been updated accordingly so that `xios_define_calendar` can accept the new attributes necessary to define custom calendars.

git-svn-id: svn+ssh://forge.ipsl.jussieu.fr/ipsl/forge/projets/ioserver/svn/XIOS@550 d613d254-d18a-4304-b296-5125d4c4d670