Support creating the timeseries automatically.
Fields which should be outputted as timeseries must have the `ts_enabled` attribute set to `true`.

Files have a new attribute `timeseries` which must be one of:
 - `none`: no timeseries are outputted, only the regular file (default behavior when the attribute is ommited).
 - `only`: only the timeseries are outputted, the regular file is not created.
 - `both`: both the timeseries and the regular files are outputted.
 - `exclusive`: the timeseries are outputted and a regular file is created with only the fields which were not marked for output as a timeserie (if any).

The name of the files created for the timeseries is composed of a prefix followed by the name of the variable. The file attribute `ts_prefix` can be used to control the prefix used (by default, the name of the regular file is used).

All the attributes of the regular file are inherited by the files created for the timeseries. The field attribute `ts_split_freq` can be used to configure the splitting for each timeseries (by default, the splitting frequency of the regular file is used).

git-svn-id: svn+ssh://forge.ipsl.jussieu.fr/ipsl/forge/projets/ioserver/svn/XIOS@711 d613d254-d18a-4304-b296-5125d4c4d670