Field: Handle more correctly the output name for the fields with a field_ref.
If the field has an explicitly defined name (which might be inherited) then it is used as the output name.
If no name was defined but an id was set, the id is used as the output name.
If no name was defined and the id was automatically generated, the id of the field directly referenced is used as the output name.

git-svn-id: svn+ssh://forge.ipsl.jussieu.fr/ipsl/forge/projets/ioserver/svn/XIOS@770 d613d254-d18a-4304-b296-5125d4c4d670